+++ docs/changelog.rst
@@ -35,6 +35,7 @@ fixes
 
 - Change the mode to entry forecast_date in runfile, this fixed problems with
   format entries that are auto-convert from sheets application
+- Fixed input validation for var I with particular values
 
 
 0.5.2a (**2013-02-13**)